+++ frontend/src/app/components/admin-create/admin-create.component.ts
@@ -2,10 +2,11 @@ import { Component, inject } from '@angular/core';
 import { AdminService } from '../../services/admin.service';
 import { Router } from '@angular/router';
 import { FormsModule } from '@angular/forms';
+import { NgIf } from '@angular/common';
 
 @Component({
   selector: 'app-admin-create',
-  imports: [FormsModule],
+  imports: [FormsModule, NgIf],
   templateUrl: './admin-create.component.html',
   styleUrl: './admin-create.component.css'
 })
@@ -17,6 +18,9 @@ export class AdminCreateComponent {
   password = '';
   code = '';
 
+  errorMessage='';
+  successMessage='';
+
   createAdmin() {
     if (!this.username || !this.password || !this.code) {
       alert('Minden mező kitöltése kötelező!');
@@ -25,12 +29,16 @@ export class AdminCreateComponent {
 
     this.adminService.createAdmin(this.username, this.password, this.code).subscribe({
       next: () => {
-        alert('Admin sikeresen létrehozva!');
-        this.router.navigate(['admin-main-menu']);
+        this.username = '';
+        this.password = '';
+        this.code = '';
+        this.errorMessage = '';
+        this.successMessage = 'Admin sikeresen létrehozva!';
       },
       error: (err) => {
         console.error('Hiba:', err);
-        alert(err.error.message);
+        this.successMessage = '';
+        this.errorMessage = err.error.message;
       }
     });
   }